Receive all UDP messages on INADDR_ANY
INADDR_ANY should receive all packets sent to the corresponding port.
The only known user of this feature is a DHCP client, so dispatch on
the UDP level.

While debugging Fuchsia's DHCP client I realize UDP broadcasts are not
delievered to INADDR_ANY. This change fixed the problem.

This doesn't seem to affect the behavior on the syscall level, added
a syscall test to demonstrate that.

PiperOrigin-RevId: 300605100